+++ ChangeLog.md
@@ -1,3 +1,30 @@
+#Ninkology 1.1 (4-14-2014)
+========
+###Bug Fixes: 
+- Archives are no longer "scanned" by the system (empty archives now give no results).<br/>
+- Now checks if file exists before attempting to scan it.<br/>
+
+###Features: 
+- If only Ninka or FOSSology is used, the licence(s) found by the single scan are "declared".<br/>
+
+#### Command Line
+- Removed '-fn' flag. Dual scan is the default behavior.<br/>
+- If both '-f' and '-n' are used, the results will be identical to not using either.<br/>
+- File information is now printed to STDOUT while the program is running to let the user know that progress is being made.<br/>
+
+#### Web Interface
+- Web Interface no longer sends e-mail to end user. Now provides the user with a link to download the scan results.<br/>
+
+###Performance improvements: 
+None
+
+###Documentation: 
+- Updated the use cases to more accurately reflect the current working version of Ninkology.<br/>
+
+
+
+-----------------------------------------------------------------------
+
 #Ninkology 1.0 (03-21-2014)
 ========
 ###Bug Fixes: